Revert "[SR-4378] extract common code of aggregate algo in order to reuse (#15)"
This reverts commit a78bf3901e9a28c2d67b580e71316f99b7a83092.
